+++ src/components/Legend/index.ts
@@ -1,2 +1,0 @@
-export {Legend} from './Legend';
-export type {Props as LegendProps} from './Legend';