+++ src/components/LegendContainer/components/Legend/index.ts
@@ -1,0 +1,2 @@
+export {Legend} from './Legend';
+export type {LegendProps} from './Legend';