+++ src/components/LegendContainer/types.ts
@@ -1,0 +1,10 @@
+import type {Color} from '../../types';
+
+export type LegendIconType = 'solid' | 'line';
+
+export interface LegendData {
+  name: string;
+  color: Color;
+  isComparison?: boolean;
+  iconType?: LegendIconType;
+}