+++ src/components/LineChart/components/Line/Line.tsx
@@ -3,13 +3,22 @@ import type {Line as D3Line} from 'd3-shape';
 import {useSpring, animated} from '@react-spring/web';
 
 import type {DataPoint, DataSeries, LineStyle} from '../../../../types';
-import {useTheme} from '../../../../hooks';
+import {
+  getColorVisionEventAttrs,
+  getOpacityStylesForActive,
+  useTheme,
+} from '../../../../hooks';
 import {ANIMATION_DELAY} from '../../constants';
-import {LINES_LOAD_ANIMATION_CONFIG} from '../../../../constants';
+import {
+  COLOR_VISION_SINGLE_ITEM,
+  LINES_LOAD_ANIMATION_CONFIG,
+} from '../../../../constants';
 
 import {StrokeDasharray} from './constants';
+import styles from './Line.scss';
 
 export interface Props {
+  activeLineIndex: number;
   series: DataSeries;
   isAnimated: boolean;
   index: number;
@@ -21,6 +30,7 @@ export interface Props {
 }
 
 export const Line = React.memo(function Shape({
+  activeLineIndex,
   series,
   isAnimated,
   index,
@@ -35,7 +45,7 @@ export const Line = React.memo(function Shape({
   const isSolidLine = series.isComparison !== true;
   const lineStyle: LineStyle = series.isComparison ? 'dotted' : 'solid';
 
-  const styles = useSpring({
+  const spring = useSpring({
     from: isSolidLine ? {transform: 'scaleY(0)'} : {opacity: 0},
     to: isSolidLine ? {transform: 'scaleY(1)'} : {opacity: 1},
     delay: isSolidLine ? index * ANIMATION_DELAY : 0,
@@ -49,9 +59,8 @@ export const Line = React.memo(function Shape({
 
   return (
     <animated.g
-      style={
-        isAnimated ? {...styles, transformOrigin: 'bottom center'} : undefined
-      }
+      className={styles.Group}
+      style={isAnimated ? {...spring, transformOrigin: 'bottom center'} : {}}
     >
       <path
         d={path}
@@ -62,8 +71,27 @@ export const Line = React.memo(function Shape({
         strokeLinejoin="round"
         strokeLinecap="round"
         strokeDasharray={StrokeDasharray[lineStyle]}
+        style={getOpacityStylesForActive({
+          activeIndex: activeLineIndex,
+          index,
+        })}
       />
       {children}
+      <path
+        className={styles.Line}
+        d={path}
+        strokeWidth={`${10}px`}
+        paintOrder="stroke"
+        stroke="transparent"
+        fill="none"
+        {...getColorVisionEventAttrs({
+          type: COLOR_VISION_SINGLE_ITEM,
+          index,
+        })}
+        style={{pointerEvents: 'auto'}}
+        aria-hidden="true"
+        tabIndex={-1}
+      />
     </animated.g>
   );
 });