+++ src/components/VerticalBarChart/components/BarGroup/tests/BarGroup.test.tsx
@@ -14,6 +14,30 @@ jest.mock('d3-scale', () => ({
 
 describe('<BarGroup/>', () => {
   const mockProps = {
+    accessibilityData: [
+      {
+        title: 'Monday',
+        data: [
+          {
+            label: 'Breakfast',
+            value: '10',
+          },
+          {
+            label: 'Lunch',
+            value: '20',
+          },
+          {
+            label: 'Dinner',
+            value: '0',
+          },
+          {
+            label: 'Snack',
+            value: '1',
+          },
+        ],
+      },
+    ],
+    activeBarGroup: -1,
     x: 10,
     yScale: scaleLinear() as any,
     width: 100,
@@ -28,6 +52,7 @@ describe('<BarGroup/>', () => {
     isSubdued: false,
     zeroAsMinHeight: false,
     isAnimated: false,
+    gapWidth: 10,
   };
 
   it('renders a <Bar /> for each data item', () => {
@@ -40,7 +65,7 @@ describe('<BarGroup/>', () => {
     expect(barGroup).toContainReactComponentTimes(Bar, 4);
   });
 
-  it("gives each <Bar /> a width that together totals the group's full width, minus spacing", () => {
+  it("gives each rect a width that together totals the group's full width, minus spacing", () => {
     const barGroup = mount(
       <svg>
         <BarGroup {...mockProps} />,
@@ -48,7 +73,7 @@ describe('<BarGroup/>', () => {
     );
 
     const bars = barGroup
-      .findAll(Bar)
+      .findAll('rect')
       .filter(
         ({props}) =>
           props.width === mockProps.width / mockProps.data.length - BAR_SPACING,
@@ -57,17 +82,17 @@ describe('<BarGroup/>', () => {
     expect(bars).toHaveLength(4);
   });
 
-  it('gives each <Bar /> a spaced out X position', () => {
+  it('gives each rect a spaced out X position', () => {
     const barGroup = mount(
       <svg>
         <BarGroup {...mockProps} />,
       </svg>,
     );
 
-    expect(barGroup).toContainReactComponent(Bar, {x: 10});
-    expect(barGroup).toContainReactComponent(Bar, {x: 35});
-    expect(barGroup).toContainReactComponent(Bar, {x: 60});
-    expect(barGroup).toContainReactComponent(Bar, {x: 85});
+    expect(barGroup).toContainReactComponent('rect', {x: 10});
+    expect(barGroup).toContainReactComponent('rect', {x: 35});
+    expect(barGroup).toContainReactComponent('rect', {x: 60});
+    expect(barGroup).toContainReactComponent('rect', {x: 85});
   });
 
   describe('zeroAsMinHeight', () => {