+++ src/components/VerticalBarChart/components/StackedBarGroups/index.ts
@@ -1,0 +1,1 @@
+export {StackedBarGroups} from './StackedBarGroups';