+++ src/components/VerticalBarChart/hooks/index.ts
@@ -1,3 +1,4 @@
 export {useXScale} from './use-x-scale';
 export {useYScale} from './use-y-scale';
 export {useMinimalLabelIndexes} from './use-minimal-label-indexes';
+export {useStackedGapsForVerticalChart} from './useStackedGapsForVerticalChart';