+++ src/components/shared/HorizontalBars/HorizontalBars.tsx
@@ -1,8 +1,9 @@
-import React from 'react';
+import React, {useState} from 'react';
 import type {ScaleLinear} from 'd3-scale';
 
 import {DataSeries, RoundedBorder} from '../../../types';
 import {
+  COLOR_VISION_SINGLE_ITEM,
   FONT_SIZE,
   FONT_SIZE_PADDING,
   HORIZONTAL_BAR_LABEL_OFFSET,
@@ -11,14 +12,18 @@ import {
 } from '../../../constants';
 import type {LabelFormatter} from '../../../types';
 import {getTextWidth} from '../../../utilities';
-import {useTheme} from '../../../hooks';
+import {
+  useTheme,
+  useWatchColorVisionEvents,
+  getColorVisionEventAttrs,
+} from '../../../hooks';
 import {Bar} from '../Bar';
 import {getGradientDefId} from '../GradientDefs';
 
 import {Label} from './components';
 
 export interface HorizontalBarsProps {
-  ariaLabel: string;
+  activeGroupIndex: number;
   barHeight: number;
   data: DataSeries[];
   groupIndex: number;
@@ -34,8 +39,8 @@ export interface HorizontalBarsProps {
 }
 
 export function HorizontalBars({
+  activeGroupIndex,
   animationDelay,
-  ariaLabel,
   barHeight,
   data,
   groupIndex,
@@ -50,11 +55,21 @@ export function HorizontalBars({
 }: HorizontalBarsProps) {
   const selectedTheme = useTheme(theme);
 
+  const [activeBarIndex, setActiveBarIndex] = useState(-1);
+
+  useWatchColorVisionEvents({
+    type: COLOR_VISION_SINGLE_ITEM,
+    onIndexChange: ({detail}) => {
+      if (activeGroupIndex === -1 || activeGroupIndex === groupIndex) {
+        setActiveBarIndex(detail.index);
+      }
+    },
+  });
+
   return (
     <g
       transform={`translate(${zeroPosition},${HORIZONTAL_GROUP_LABEL_HEIGHT})`}
-      aria-label={ariaLabel}
-      role="listitem"
+      aria-hidden="true"
     >
       {data.map((_, seriesIndex) => {
         if (data[seriesIndex].data[groupIndex] == null) {
@@ -64,6 +79,9 @@ export function HorizontalBars({
 
         const isNegative = value && value < 0;
         const label = labelFormatter(value);
+        const ariaLabel = `${
+          data[seriesIndex] ? data[seriesIndex].name : ''
+        } ${value}`;
 
         const labelWidth = getTextWidth({
           text: `${label}`,
@@ -80,7 +98,6 @@ export function HorizontalBars({
           HORIZONTAL_SPACE_BETWEEN_SINGLE * seriesIndex;
         const negativeX = (width + leftLabelOffset) * -1;
         const x = isNegative ? negativeX : width + HORIZONTAL_BAR_LABEL_OFFSET;
-        const ariaHidden = seriesIndex !== 0;
 
         return (
           <React.Fragment key={`series-${seriesIndex}-${id}-${name}`}>
@@ -89,11 +106,10 @@ export function HorizontalBars({
               color={`url(#${getGradientDefId(theme, seriesIndex, id)})`}
               height={barHeight}
               index={groupIndex}
+              isActive={activeBarIndex === -1 || activeBarIndex === seriesIndex}
               isAnimated={isAnimated}
               needsMinWidth
-              role="img"
               roundedBorder={RoundedBorder.Right}
-              tabIndex={ariaHidden ? -1 : 0}
               transform={isNegative ? 'scaleX(-1)' : ''}
               width={width}
               x={0}
@@ -115,6 +131,21 @@ export function HorizontalBars({
                 y={y}
               />
             )}
+            <rect
+              x={0}
+              y={y - HORIZONTAL_SPACE_BETWEEN_SINGLE / 2}
+              width={width}
+              height={barHeight + HORIZONTAL_SPACE_BETWEEN_SINGLE}
+              fill="transparent"
+              style={{transform: isNegative ? 'scaleX(-1)' : ''}}
+              {...getColorVisionEventAttrs({
+                type: COLOR_VISION_SINGLE_ITEM,
+                index: seriesIndex,
+              })}
+              tabIndex={-1}
+              role="img"
+              aria-label={ariaLabel}
+            />
           </React.Fragment>
         );
       })}