+++ src/components/shared/HorizontalStackedBars/HorizontalStackedBars.tsx
@@ -1,17 +1,15 @@
-import React, {useMemo} from 'react';
+import React, {useMemo, useState} from 'react';
 import type {ScaleLinear} from 'd3-scale';
 import {animated, useSpring} from '@react-spring/web';
 
+import {useWatchColorVisionEvents} from '../../../hooks';
 import {getBarId} from '../../../utilities';
 import {
   BARS_TRANSITION_CONFIG,
+  COLOR_VISION_SINGLE_ITEM,
   HORIZONTAL_GROUP_LABEL_HEIGHT,
 } from '../../../constants';
-import {
-  DataSeries,
-  FormattedStackedSeries,
-  RoundedBorder,
-} from '../../../types';
+import {FormattedStackedSeries, RoundedBorder} from '../../../types';
 import {getGradientDefId} from '..';
 
 import {StackedBar} from './components';
@@ -19,15 +17,16 @@ import {useStackedGaps} from './hooks';
 import {getXPosition} from './utilities';
 
 export interface HorizontalStackedBarsProps {
+  activeGroupIndex: number;
   animationDelay: number;
   ariaLabel: string;
   barHeight: number;
-  data: DataSeries[];
+  dataKeys: string[];
   groupIndex: number;
   id: string;
   isAnimated: boolean;
   name: string;
-  stackedValues: FormattedStackedSeries;
+  stackedValues: FormattedStackedSeries[];
   xScale: ScaleLinear<number, number>;
   theme?: string;
 }
@@ -53,9 +52,10 @@ function getRoundedBorder({
 }
 
 export function HorizontalStackedBars({
+  activeGroupIndex,
   animationDelay,
-  ariaLabel,
   barHeight,
+  dataKeys,
   groupIndex,
   id,
   isAnimated,
@@ -64,6 +64,17 @@ export function HorizontalStackedBars({
   theme,
   xScale,
 }: HorizontalStackedBarsProps) {
+  const [activeBarIndex, setActiveBarIndex] = useState(-1);
+
+  useWatchColorVisionEvents({
+    type: COLOR_VISION_SINGLE_ITEM,
+    onIndexChange: ({detail}) => {
+      if (activeGroupIndex === -1 || activeGroupIndex === groupIndex) {
+        setActiveBarIndex(detail.index);
+      }
+    },
+  });
+
   const {transform} = useSpring({
     from: {
       transform: `scale(0, 1) translate(0, ${HORIZONTAL_GROUP_LABEL_HEIGHT}px`,
@@ -97,11 +108,14 @@ export function HorizontalStackedBars({
 
   return (
     <animated.g
-      aria-label={ariaLabel}
-      role="listitem"
-      style={{transform, transformOrigin: `${xScale(0)}px 0px`}}
+      style={{
+        transform,
+        transformOrigin: `${xScale(0)}px 0px`,
+      }}
+      aria-hidden="true"
     >
-      {stackedValues[groupIndex].map(([start, end], seriesIndex) => {
+      {stackedValues[groupIndex].map((item, seriesIndex) => {
+        const [start, end] = item;
         const barId = getBarId(id, groupIndex, seriesIndex);
         const width = Math.abs(xScale(end) - xScale(start));
 
@@ -115,18 +129,22 @@ export function HorizontalStackedBars({
         });
 
         const x = getXPosition({start, end, seriesIndex, gaps, xScale});
+        const key = dataKeys[seriesIndex] ?? '';
+        const ariaLabel = `${key} ${end}`;
 
         return (
           <StackedBar
+            activeBarIndex={activeBarIndex}
+            ariaLabel={ariaLabel}
             color={getGradientDefId(theme, seriesIndex, id)}
-            groupIndex={groupIndex}
             height={barHeight}
             isAnimated={isAnimated}
             key={`${name}${barId}`}
+            roundedBorder={roundedBorder}
             seriesIndex={seriesIndex}
+            setActiveBarIndex={setActiveBarIndex}
             width={width}
             x={x}
-            roundedBorder={roundedBorder}
           />
         );
       })}