+++ src/hooks/index.ts
@@ -21,3 +21,10 @@ export {useHorizontalSeriesColors} from './useHorizontalSeriesColors';
 export type {HorizontalTransitionStyle} from './useHorizontalTransitions';
 export {useBarChartTooltipContent} from './useBarChartTooltipContent';
 export {useHorizontalStackedValues} from './useHorizontalStackedValues';
+export {
+  useColorVisionEvents,
+  getColorVisionEventAttrs,
+  useWatchColorVisionEvents,
+  getOpacityStylesForActive,
+} from './ColorVisionA11y';
+export {useCallbackRef} from './useCallbackRef';